Define `VUE_APP_WS_ADDRESS` from env variable
Primarily this makes it possible to easily redefine server that was
hardcoded in `Vue-IFrame.ts` and `manifest.json`.

As a side effect it is now possible to define `content_security_policy`
based on mode. Production builds no longer include `connect-src` to
staging server and other policies that used to be there for
`vue-remote-devtools`.